Rename: `to_ordered_queue` -> `to_ordered_list`
