+++ resources/ddl/define-all-as-permanent.hive
@@ -464,6 +464,13 @@ DROP FUNCTION IF EXISTS to_ordered_map;
 CREATE FUNCTION to_ordered_map as 'hivemall.tools.map.UDAFToOrderedMap' USING JAR '${hivemall_jar}';
 
 ---------------------
+-- list functions --
+---------------------
+
+DROP FUNCTION IF EXISTS to_ordered_list;
+CREATE FUNCTION to_ordered_list as 'hivemall.tools.list.UDAFToOrderedList' USING JAR '${hivemall_jar}';
+
+---------------------
 -- Math functions --
 ---------------------
 
@@ -517,9 +524,6 @@ CREATE FUNCTION x_rank as 'hivemall.tools.RankSequenceUDF' USING JAR '${hivemall
 
 DROP FUNCTION IF EXISTS each_top_k;
 CREATE FUNCTION each_top_k as 'hivemall.tools.EachTopKUDTF' USING JAR '${hivemall_jar}';
-
-DROP FUNCTION IF EXISTS to_ordered_queue;
-CREATE FUNCTION to_ordered_queue as 'hivemall.tools.UDAFToOrderedQueue' USING JAR '${hivemall_jar}';
 
 -------------------------------
 -- Text processing functions --
