+++ resources/ddl/define-all.hive
@@ -456,6 +456,13 @@ drop temporary function if exists to_ordered_map;
 create temporary function to_ordered_map as 'hivemall.tools.map.UDAFToOrderedMap';
 
 ---------------------
+-- list functions --
+---------------------
+
+drop temporary function if exists to_ordered_list;
+create temporary function to_ordered_list as 'hivemall.tools.list.UDAFToOrderedList';
+
+---------------------
 -- Math functions --
 ---------------------
 
@@ -509,9 +516,6 @@ create temporary function x_rank as 'hivemall.tools.RankSequenceUDF';
 
 drop temporary function if exists each_top_k;
 create temporary function each_top_k as 'hivemall.tools.EachTopKUDTF';
-
-drop temporary function if exists to_ordered_queue;
-create temporary function to_ordered_queue as 'hivemall.tools.UDAFToOrderedQueue';
 
 -------------------------------
 -- Text processing functions --
